# swagger.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "OpenAPI definition",
    "version": "v0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Generated server url"
    }
  ],
  "paths": {
...
    "/api/locations/{id}": {
      "get": {
        "tags": [
          "location-controller"
        ],
        "operationId": "getLocationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        }
      },
      "put": {
        "tags": [
          "location-controller"
        ],
        "operationId": "updateLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Location"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "location-controller"
        ],
        "operationId": "deleteLocation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/locations": {
      "get": {
        "tags": [
          "location-controller"
        ],
        "operationId": "getAllLocations",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Location"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "location-controller"
        ],
        "operationId": "createLocation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Location"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Location"
                }
              }
            }
          }
        }
      }
    },
    "/api/legal-info": {
      "get": {
        "tags": [
          "legal-information-controller"
        ],
        "operationId": "getAllLegalInformation",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/LegalInformation"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "legal-information-controller"
        ],
        "operationId": "createLegalInformation",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LegalInformation"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LegalInformation"
                }
              }
            }
          }
        }
      }
    },
    "/api/landmarks": {
      "get": {
        "tags": [
          "nearby-landmarks-controller"
        ],
        "operationId": "getAllLandmarks",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/NearbyLandmarks"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "nearby-landmarks-controller"
        ],
        "operationId": "createLandmark",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/NearbyLandmarks"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyLandmarks"
                }
              }
            }
          }
        }
      }
    },
    "/api/amenities": {
      "get": {
        "tags": [
          "amenities-controller"
        ],
        "operationId": "getAllAmenities",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Amenities"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "amenities-controller"
        ],
        "operationId": "createAmenity",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/Amenities"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Amenities"
                }
              }
            }
          }
        }
      }
    },
    "/api/additional-details": {
      "get": {
        "tags": [
          "additional-details-controller"
        ],
        "operationId": "getAllAdditionalDetails",
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdditionalDetails"
                  }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "additional-details-controller"
        ],
        "operationId": "createAdditionalDetails",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AdditionalDetails"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdditionalDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/legal-info/{id}": {
      "get": {
        "tags": [
          "legal-information-controller"
        ],
        "operationId": "getLegalInformationById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/LegalInformation"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "legal-information-controller"
        ],
        "operationId": "deleteLegalInformation",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/landmarks/{id}": {
      "get": {
        "tags": [
          "nearby-landmarks-controller"
        ],
        "operationId": "getLandmarkById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/NearbyLandmarks"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "nearby-landmarks-controller"
        ],
        "operationId": "deleteLandmark",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/amenities/{id}": {
      "get": {
        "tags": [
          "amenities-controller"
        ],
        "operationId": "getAmenityById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/Amenities"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "amenities-controller"
        ],
        "operationId": "deleteAmenity",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/additional-details/{id}": {
      "get": {
        "tags": [
          "additional-details-controller"
        ],
        "operationId": "getAdditionalDetailsById",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "*/*": {
                "schema": {
                  "$ref": "#/components/schemas/AdditionalDetails"
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "additional-details-controller"
        ],
        "operationId": "deleteAdditionalDetails",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "Location": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "city": {
            "type": "string"
          },
          "state": {
            "type": "string"
          },
          "address": {
            "type": "string"
          },
          "latitude": {
            "type": "number",
            "format": "double"
          },
          "longitude": {
            "type": "number",
            "format": "double"
          }
        }
      },
      "LegalInformation": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "reraNumber": {
            "type": "string"
          },
          "approvals": {
            "type": "string"
          }
        }
      },
      "NearbyLandmarks": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "school": {
            "type": "string"
          },
          "hospital": {
            "type": "string"
          },
          "shoppingMall": {
            "type": "string"
          },
          "metroStation": {
            "type": "string"
          },
          "airport": {
            "type": "string"
          }
        }
      },
      "Amenities": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "name": {
            "type": "string"
          }
        }
      },
      "AdditionalDetails": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int64"
          },
          "furnishedStatus": {
            "type": "string"
          },
          "greenRating": {
            "type": "string"
          },
          "petFriendly": {
            "type": "boolean"
          },
          "carParking": {
            "type": "string"
          },
          "totalCarParkingSpaces": {
            "type": "integer",
            "format": "int32"
          }
        }
      }
    }
  }
}